Move tech LEF metals parsing to LEFUtils (a bit clunky), fix type checker errors
