+++ hammer/par/openroad/__init__.py
@@ -1450,10 +1450,7 @@ class OpenROADPlaceAndRoute(OpenROADPlaceAndRouteTool):
 
         layer_map_file=self.get_gds_map_file()
         if layer_map_file is None:
-            self.logger.error("Must have GDS layer map for macro PG library generation! Skipping.")
-            return True
-        else:
-            assert isinstance(layer_map_file, str)
+            raise FileNotFoundError(f"Must have GDS layer map for write_gds! KLayout won't be able to write GDS file. ({layer_map_file}")
 
         # the first entry of $KLAYOUT_PATH will be the one where the configuration is stored when KLayout exits
         #   otherwise KLayout tries to write everything to the same directory as the klayout binary and throws an error if it is not writeable
